Fix binding configs to switch IME
As a workaround for a mutter issue where IME switching is triggered twice
with a single keypress, configure `locate-pointer-key='Control_R'` and
remove the `Alt + Space` keyboard shortcut.

The issue occurs when `Control + Space` is bound to switch IME and
`locate-pointer-key` is set to its default `Control_L`. The problem only
manifests with the left Control key, suggesting a conflict with the default
locate-pointer configuration. Setting `locate-pointer-key` to `Control_R`
avoids this conflict, as the right Control key is rarely used.

Link: https://gitlab.gnome.org/GNOME/mutter/-/issues/4404#note_2578079
Link: https://github.com/ibus/ibus/issues/2814